# apps/api/docs/swagger.yaml
definitions:
  interfaces.OperationListResponse:
    properties:
      processes:
        description: List of processes with their current status
        items:
          $ref: '#/definitions/interfaces.OperationReview'
        type: array
    type: object
  interfaces.OperationProgress:
    properties:
      percentage:
        type: integer
      processed_files:
        type: integer
      total_files:
        type: integer
    type: object
  interfaces.OperationResults:
    properties:
      files_processed:
        items:
          type: string
        type: array
      most_frequent_words:
        items:
          type: string
        type: array
      total_lines:
        type: integer
      total_words:
        type: integer
    type: object
  interfaces.OperationReview:
    properties:
      error:
        description: Error message if the process failed
        type: string
      estimated_completion:
        description: Estimated completion time
        type: string
      id:
        description: Unique identifier for the process
        type: string
      started_at:
        description: Timestamp when the process started
        type: string
      status:
        description: Current status of the process (e.g., "pending", "in_progress",
          "completed", "failed")
        type: string
    type: object
  interfaces.OperationStatus:
    properties:
      completed_at:
        description: Timestamp when the process completed
        type: string
      error:
        description: Error message if the process failed
        type: string
      estimated_completion:
        description: Time taken for the process to complete
        type: string
      process_id:
        description: Unique identifier for the process
        type: string
      progress:
        allOf:
        - $ref: '#/definitions/interfaces.OperationProgress'
        description: Progress details of the process
      results:
        allOf:
        - $ref: '#/definitions/interfaces.OperationResults'
        description: Result of the process (e.g., summary, statistics)
      started_at:
        description: Timestamp when the process started
        type: string
      status:
        allOf:
        - $ref: '#/definitions/interfaces.OperationStatusEnum'
        description: Current status of the process
    type: object
  interfaces.OperationStatusEnum:
    enum:
    - PENDING
    - RUNNING
    - PAUSED
    - COMPLETED
    - FAILED
    - STOPPED
    type: string
    x-enum-varnames:
    - Pending
    - Running
    - Paused
    - Completed
    - Failed
    - Stopped
  interfaces.SummarizeHTTPOutput:
    properties:
      file_name:
        type: string
      file_size:
        type: integer
      summary:
        type: string
    type: object
  interfaces.SummarizeInput:
    properties:
      file_path:
        type: string
      max_length:
        type: integer
    type: object
info:
  contact: {}
paths:
  /process/list:
    get:
      consumes:
      - application/json
      description: List all processes with their current status
      produces:
      - application/json
      responses:
        "200":
          description: List of Processes
          schema:
            $ref: '#/definitions/interfaces.OperationListResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: List Processes
      tags:
      - Process
  /process/results/{id}:
    get:
      consumes:
      - application/json
      description: Get the results of a process by its ID
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Process Results
          schema:
            $ref: '#/definitions/interfaces.OperationStatus'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Process Not Found
          schema:
            type: string
      summary: Get Process Results
      tags:
      - Process
  /process/start:
    post:
      consumes:
      - application/json
      description: Start a new process with the provided data
      parameters:
      - description: Process Data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Process Started
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Start a Process
      tags:
      - Process
  /process/status/{id}:
    get:
      consumes:
      - application/json
      description: Get the status of a process by its ID
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Process Status
          schema:
            $ref: '#/definitions/interfaces.OperationStatus'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Process Not Found
          schema:
            type: string
      summary: Get Process Status
      tags:
      - Process
  /process/stop/{id}:
    post:
      consumes:
      - application/json
      description: Stop a process by its ID
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Process Status
          schema:
            $ref: '#/definitions/interfaces.OperationStatus'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Process Not Found
          schema:
            type: string
      summary: Stop a Process
      tags:
      - Process
  /summarizer/mcp:
    post:
      consumes:
      - application/json
      description: Handle MCP requests for summarization
      parameters:
      - description: Summarize Input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/interfaces.SummarizeInput'
      produces:
      - application/json
      responses:
        "200":
          description: Summarize Output
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: MCP Handler for Summarizer
      tags:
      - Summarizer
  /summarizer/summarize:
    post:
      consumes:
      - application/json
      description: Handle http requests for summarization
      parameters:
      - description: Summarize Input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/interfaces.SummarizeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.SummarizeHTTPOutput'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: http Handler for Summarizer
      tags:
      - Summarizer
swagger: "2.0"